# Measure performance of exp in G, N, and N^2