--- a/tests/benches/exp.rs
+++ b/tests/benches/exp.rs
@@ -1,5 +1,3 @@
-use std::hint::black_box;
-
 use cggmp21::{
     fast_paillier::utils::external_rand,
     rug::{self, Complete},