In the paper, authors give estimation of performance of each protocol based on cost of point mult, and exp in $N$, $N^2$.

![image](https://github.com/user-attachments/assets/3c42a217-40dd-4182-864d-abe924cb78c1)

From my benchmarks on my computer:
* Point multiplication takes 50.347 µs
* $x^e \bmod N$
  * $\left | e \right | = \left | q \right | = 256$, 508.87 µs
  * $\left | e \right | = \ell + \varepsilon = 486$, 945.65 µs
  * $\left | e \right | = \ell' + \varepsilon = 1078$, 2.0480 ms
* $x^e \bmod N^2$
  * $\left | e \right | = \ell = 256$, 1.8894 ms
  * $\left | e \right | = \ell + \varepsilon = 486$, 3.5165 ms
  * $\left | e \right | = \left | N \right | = 3072$, 21.270 ms

The paper says that estimation of presigning perf is:

$$n \cdot (12G + 56N + 33N^2) = 2.4 \text{sec}$$

(if we take the highest cost per $N$ and $N^2$)

The actual perf of presigning on my computer is 2.71s which is pretty close. 

It's a bit strange that actual perf is higher than the estimation: I expected the estimation to be an upper bound of the performance as not always we do exponentiation with the highest exponent size, sometimes it's optimized with CRT, etc. Probably, perf estimation doesn't take into account some of the computations.